Added the start of a new public web, based on the sourceforge site's design.
There is still lots to do. The first step is to get the full functionality of the old web into the new format. Step two will be to get rid of all or most of the java code in the jsp pages.

To try out the new web, build and deploy as usual, then browse to

http://localhost:8080/new/ejbca/

The old web is still in place. If you do not want the new web on your site, remove the dependency from ca.ear to publicweb.jar.


git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@4207 37aeb1de-b14b-0410-9a0e-82b28170d228