ECA-1806:
Redesigned the timers a bit for this ticket. In EJB3, timers are assigned to the bean that created them, which meant that the relationship between ServiceTimerSessionBean and LocalerviceSessionBean was faulty, having timer creation and timer callback shared between them. Moving the creation methods into ServiceTimerSession created a circular dependency, and on analysis the two beans can't exist independent of each other and perform the same tasks, so the methods from ServiceTimerSessionBean have been merged into LocalServiceSessionBean and the bean has been removed.
Additionally, the timer removal methods in LocalServiceSessionBean have been made more granular, in order to avoid transactional problems when writing to both hypersonic and the user database in the same transaction.

CertificateExpiration has been modified with a further safety, stopping it from crashing if no CAs are in existence. 

CertificateExpireTest was modified slightly in preparation for JUnit4 being introduced. 

git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@9780 37aeb1de-b14b-0410-9a0e-82b28170d228