ECA-11077: Remove the CMS service
Removed the CMS service from the CA settings and audit log page, and the
extended key services key spec, which was only used for the CMS service.
Not bothering with a cleanup job in the upgrade, because it would have
little impact while it's always a bit risk removing data.