ECA-7618: Provide separate field for SCP publisher port
This is to address potential Windows-style paths issues, where drives
are named with a letter followed by colons. It virtually resolves
previous path parsing, but does not really resolve in full extent.
Notable case is local path (without host) with a Windows drive in it.
This case is ignored as it was previously, just not introducing new
issues. :)