ECA-11819: Avoid database lookup for log redaction
In createPKCS7 there was db lookup to get the EEP for the certificate.
We suspect this db lookup might be the cause of a NPE in certain cases,
if the certificate that was just issued has not been persisted.
This commit removes the DB lookup by passing the EEP id as an input
parameter to the method.