Edit End Entity: reorder buttons, so Revoke And Delete is not the default action
<h:commandButton> gets rendered as <input type=submit>. Pressing the
Enter key in an input field on an form with multiple submit buttons is
equivalent to clicking the first one. Up to now, the default action of
the Edit End Entity form was the dangerous operation "Revoke
certificates and delete end entity." It is probably more likely that the
user intended to save changes instead, which is the action that is now
the default (executed when pressing Enter).