ECA-13070: Remove node selectors
The node selectors were added to allow EJBCA cli's to be used in an HA
environment that is configured with shared EJBs.  Future experimentation
showed that this is not necessary for EJBCA - EJBCA HA works best with
shared sessions, but exclusively local EJBs.

Also, they require supporting classes from the app server which may not
be there, depending on the app server configuration.  In particular, the
Jenkins build is building against an app server which doesn't have the
supporting classes.

Remove them and the common module that was added just to hold them